test: fix type errors
