+++ plugins/hitokoto/__tests__/hitokoto.spec.ts
@@ -3,7 +3,7 @@ import { App } from 'koishi'
 import mock from '@koishijs/plugin-mock'
 import memory from '@koishijs/plugin-database-memory'
 
-import * as hitokoto from '..'
+import * as hitokoto from '../src'
 
 describe('basic', () => {
   const app = new App()
