+++ plugins/hitokoto/package.json
@@ -37,7 +37,7 @@
     "build": "constructeur build && tsc --emitDeclarationOnly",
     "dev": "constructeur build --minify false && tsc --emitDeclarationOnly",
     "clean": "rimraf dist/*",
-    "test": "mocha -r ts-node/register --extension .spec.ts ./__tests__"
+    "test": "mocha -r ts-node/register -r yml-register --extension .spec.ts ./__tests__"
   },
   "bugs": {
     "url": "https://github.com/AwesomeHamster/koishi-plugin-hitokoto/issues"
@@ -77,6 +77,7 @@
     "mocha": "^9.1.3",
     "rimraf": "^3.0.2",
     "ts-node": "^10.8.1",
-    "typescript": "^4.7.3"
+    "typescript": "^4.7.3",
+    "yml-register": "^1.0.0"
   }
 }
