+++ plugins/macrodict/src/utils.ts
@@ -1,3 +1,31 @@
+import { Search } from './search'
+
+export type MacroDictDatabase = Record<
+  LocalizedKeys<CommandPrefix | 'Description'>,
+  string
+> & {
+  id: number
+  lastUpdated: number
+}
+
+declare module 'koishi' {
+  interface Tables {
+    macrodict: MacroDictDatabase
+  }
+
+  namespace Context {
+    interface Services {
+      macrodict: Search
+    }
+  }
+
+  interface Events {
+    /* eslint-disable @typescript-eslint/naming-convention */
+    'macrodict/update': () => void
+    /* eslint-enable @typescript-eslint/naming-convention */
+  }
+}
+
 export const locales = ['en', 'de', 'fr', 'ja', 'ko', 'chs'] as const
 export type Locale = typeof locales[number]
 export const commandPrefix = [
